requirements.txt
jmespath~=1.0.1
PyYAML~=6.0.1
pyyaml-include~=2.2
Flask~=3.1.0
Flask-SQLAlchemy~=3.1.1
Flask-Caching~=2.3.1
python-memcached
SQLAlchemy~=2.0.27
smsru-api~=1.2
huawei-lte-api~=1.8.1
gunicorn~=23.0.0
bcrypt~=4.3.0
environs~=14.5.0
pyrad2~=1.2.0